fix comparison of signed, unsigned

+++ src/readers.cpp
@@ -2,10 +2,12 @@
 // [[Rcpp::depends(RcppArmadillo)]]
 
 //' @export
+// @param X matrix of clustering solutions, rows = each solution, cols = each item to be clustered
+// @param values a vector containing all the unique values in X
 // [[Rcpp::export]]
 arma::mat calc_psm(arma::mat X, arma::vec values) {
   arma::mat PSM(X.n_cols, X.n_cols);
-  for (int i=0; i < values.n_elem; ++i) {
+  for (int i=0; (unsigned) i < values.n_elem; ++i) {
     arma::uvec idx = find(X == values(i)); // Substitute == with >, <, >=, <=, !=
     arma::mat isequal = arma::zeros(X.n_rows,X.n_cols);
     isequal.elem(idx).fill(1.0);        // Retrieve elements from positional index
